requirements.txt
colorama
requests 
python-whois 
dnspython
